connectors/requirements.txt
httpx==0.27.2
jsonschema==4.23.0
paho-mqtt==2.1.0
psycopg[binary]==3.2.6
